Fix device log bug
When we get the device log, the C++ code starts to print on the stdout and in the JavaScript code we listen for the data event. If there is enaugh time between each message which is printed on the stdout the data event is raised with only one message. If the C++ code starts to print on the stdout too fast (which is the case with the device log), NodeJS buffers all the data and emits data event with all the buffered data. This breaks out code because we expect only one message. The solution for this problem is to write method which will "parse" the buffered data. This method should check the header of the received buffered data and read message from the buffered data. If there is full message in the buffered data we will read it. If there is no full message we should store the unfinished message and on the next received buffered data we should concat the unfinished message and the new data and try to parse the concatenated message. We rely on that the first message should have header with length. After that the method will make sure that the unfinished message will always begin with header with length and that each recursive call will receive data which begins with header with length.

While testing the new method with 7 threads which print on the stdout we noticed that at some point the data which is printed on the stdout gets malformed. For example with this message:
`{"message1":"Feb 10 13:59: 28 dragon - iPhone6 timed[65] <Notice> : (Error)TimeLink : B76E090A - 9807 - 4AA3 - BB01 - 3CDE3DD48D55(4 \n- TMLSSourceDevice) send failure(Error Domain = com.apple.identityservices.error Code = 23 UserInfo=0x12fe326f0 {NSLocalizedDescription=Timed out, NSUnderlyingError=0x12fe34190 The operation couldnt be completed. (com.apple.ids.idssenderrordomain error 12.)}), will retry"}`
we receive random number of messages before we receive one message with additional `}`. With one thread we do not receive malformed output. To solve this problem we need to add mutex in the print method and make it thread safe.