Expose the socket returned from the ios runtime
The socket returned from the ios runtime can be accessed only from the ios-device-lib. We need a way to expose it.
The solution is to create a server which listnes to a port and when the client application connects to this address we will proxy the messages from the client socket to the device socket and from the device socket to the client socket.