Fix crash when disconnecting devices
Whenever disconnecting devices a query might be in progress inquiring about the device's applications. The query then may read 0 bytes of memory and due to the lack of check for this we transformed a buffer containing random memory into a number and later on allocated as much memory.
This lead to an access violation as well as a crash.