Fix error when application is uninstalled manually
In case the application is uninstalled manually during working with current device, trying to install it again fails as at the end of install we are cleaning the old resources. However the for loop is modifying the object over which we iterrate and the code fails.
So copy the object before iterrating over it.