Fix stop application failure when app is uninstalled manually from device
In case you have the following workflow
- start application on device from C++ code
- uninstall the application manually
- install the application from C++ code
- call stop_application (i.e. restart - stop-start)

The code fails as the gdb instance returns empty answers. In order to resolve the issue, clean the gdb server instance when app is installed. This way when the stop_application method is called, we'll init a new gdb and the stop will return correct results.