Don't cache the notification proxy service
When we execute awaitNotification and there is no response we will wait to
receive something from the notification proxy service socket. If we execute it
again while we are waiting the response from the first call we will start
waiting for response on the same socket. If only one response comes in the
socket one of the calls will get it and the other one will continue
waiting. To avoid this we should not cache the notification proxy service
(the socket). This way each call will start new service and wait for response
on different socket.

We had problems on some devices when we start new service for every call
before we started to use the ios-device-lib. That's why we want to disable
the cache only for the notification service proxy.

Tests with 500 postNotification invocations:
- there are no repeating sockets
- sending messages to each socket returns correct bytes sent count

There is also a workaround which will not fix the problem but it will make
it harder to reproduce. We can implement timeout in receive_message method
and this way we will wait for response only for 1 second for example. The bug
will appear only if we execute awaitNotification two times in 1 second.