fix: MaxListenersExceededWarning is thrown and breaks current process
In case iOS device is detached, the pending promises are never fulfilled. Attaching multiple devices and detaching them leads to MaxListenersExceededWarning as for the pending promises we have not detached the handlers.
```
(node:4520) MaxListenersExceededWarning: Possible EventEmitter memory leak detected. 11 data listeners added. Use emitter.setMaxListeners() to increase limit
```

Fix this by rejecting all promises in case device is detached.
Also add timeout to the options passed to IOSDeviceLib constructor. The timeout will be used for each action and in case the promise is not fulfilled until the timeout passes, it will be rejected.
The timeout can be disregarded for specific actions, currently only for device logs.