fix: Unable to assign to constant variable
In case device is detached during waiting for some action (except device logs) we should reject the promise and throw an error.
In this case we construct a string message, persist it in a constant variable and later we try to reassing the value of the constant variable.
Node.js raises an error - "TypeError: Assignment to constant variable"...
We must use let, not const here.