fix: message-unpack-stream parses messages incorrectly in some cases
The `message-unpack-stream` parses messages incorrectly in case a message with length less than Header length (4) is received - this leads to indefinite recursion. The problem has been introduced in https://github.com/telerik/ios-device-lib/pull/24/ which was fixing another issue - when the length of the new message is less than the buffer header and there's something in the `_unfinishedMessage`.
To resolve the issue, get the previous logic and improve the check - in case we have partial message from the previous time, just add the current message to it and get in the recursion. In case we do not have previous message and current data is less than 4 bytes, just put it in the `_unfinishedMessage` and wait for new data.