fix: add timeout when receiving message from connection
All socket connections were changed and now are started in ssl mode. This changes the way messages are sent and received from sockets. As a result, `timeout` option is not anymore respected when receiving messages from socket. However, the messages now are received only via methods from mobile device lib. As there isn't a built-in method in mobile device that can work with timeout, we implemented timeout mechanism using new thread. `AMDServiceConnectionReceive` blocks the current thread when application is not running on device. After the provided timeout, the socket connection is invalidated and this way the current thread is freed.