fix: clear timeout after receiving successfully message from socket's connection
A specified timeout is set when receiving a message from socket's connection. However, this timeout is never cleared. This lead to the issue that socket's connection is invalidated when receiving successfully message. After that it is not possible to use the socket's connection again in the same process.

It seems that `mobile-device` lib doesn't clear automatically connections. Starting a service with skip_cache: true can potentially lead to some random issues regarding the allowed count of started services on device.