fix: setTimeout does not work on Windows
As the pthread_cancel function is not available on Windows, change the definition of setTimeout method - from now on when you want to use it, you should pass the timeout, the data that you want to be passed to the function that will be executed in case the timeout passes and the function itself.
The setTimeout method automatically allocates memory for the objects, so they will still live even when the function from which you call the setTimeout finishes its execution.
It is not a good idea to capture variables (with lambda functions) when we'll use them in a separate thread. As we do not await the execution of the thread, the captured variables may be cleared (method from which we've called setTimeout finished its execution). More information is available here http://bannalia.blogspot.com/2016/07/passing-capturing-c-lambda-functions-as.html
NOTE: Changes are not handled for macOS.
NOTE: Allocated memory should be cleaned in clearTimeout as well.