fix: don't stuck on "Restarting application" when some error occurs while stopping the application
Currently when the application is stopped, NativeScript CLI hangs up on "Restarting application". This happens as we're no able to start `gdb server` due to some error. (for example, not mounted developer disk image on device). In this situation, `ios-device-lib` doesn't print any message back to the CLI and this way the entire `tns run` command is blocked. Printing error in this situation will ensure that NativeScript CLI will continue the command execution.

Rel to https://github.com/NativeScript/nativescript-cli/issues/5020