fix: fix the behavior of setTimeout
Currently `ios-device-lib` hangs when listing the installed applications on connected device. This happens when application is ran using sidekick. The applications are gathered from device on every 200ms. It seems that for some reasons, `ios-device-lib` is not able to create new `std::thread` and it hangs on `std::thread(operation)` on some windows machines. The result from listed applications is reported by chunks and a new `std::thread` is created for every reported chunk. As a workaround of this problem, we'll set timeout to `0` when listing the applications. (It was the default behavior before introducing `setTimeout` function). This PR also changes the interface of `setTimeout` and `clearTimeout` and makes them working with pointers instead of structs.